Refactor Registry components: rename ListAgents file and remove unused AgentDetails component
